ai/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
openai==1.106.1
pydantic==2.11.7
pydantic-settings==2.10.1
PyJWT==2.10.1
pytest==8.4.1
